+++ packages/apps/stacc-website/src/pages/Case.js
@@ -1,6 +1,5 @@
 import React from "react"
-import styled from "styled-components"
-import { Layout, theming } from "@staccx/base"
+import { Layout } from "@staccx/base"
 import BlockContent from "@sanity/block-content-to-react"
 import blockContentSerializer from "./blockContentSerializer"
 import { NavigationSubpage, SectionHead } from "../components/_codeSplitting"
@@ -14,13 +13,6 @@ const Case = ({ match, location, caseStudy }) => (
         lede={caseStudy.header.body}
         illustration={caseStudy.header.image}
       />
-      <BContenty>
-        <BlockContent
-          blocks={caseStudy.body}
-          serializers={blockContentSerializer}
-          renderContainerOnSingleChild
-        />
-      </BContenty>
       <BlockContent
         blocks={caseStudy.blocks}
         serializers={blockContentSerializer}
@@ -29,27 +21,5 @@ const Case = ({ match, location, caseStudy }) => (
     </Layout>
   </div>
 )
-
-const BContenty = styled(Layout)`
-  > p {
-    max-width: ${theming.wrapper.medium};
-    line-height: 1.6;
-  }
-
-  > blockquote {
-    padding: ${theming.spacing.large} 0;
-
-    > p {
-      font-size: ${theming.font.h3};
-      font-style: italic;
-      font-weight: 300;
-      line-height: 1.6;
-      margin-bottom: ${theming.spacing.small};
-    }
-    > cite {
-      font-style: normal;
-    }
-  }
-`
 
 export default Case